Less nested ISA (#698)
