+++ sdk/src/client/api/block_builder/input_selection/burn.rs
