+++ sdk/src/client/api/block_builder/input_selection/error.rs
