+++ sdk/src/client/api/block_builder/input_selection/requirement/alias.rs
